read/expand: drop redundant casts flagged by clippy
- variable_buffer_byte/set_variable_buffer_byte: `off` is already `usize`
  (size_t), so the `as usize` index casts were redundant.
- Reconstruct the adjusted `colonp` and the post-`parse_file_seq` `p2` via the
  original `colonp.offset(..)` pointer arithmetic instead of
  `variable_buffer.offset((usize) as isize)`, avoiding clippy's
  `ptr_offset_with_cast`. Only the dereferences are routed through the
  bounds-checked accessors; the pointer arithmetic (not a deref) is unchanged
  from the original and is not what CodeQL flags.

Behavior unchanged; clippy 354.

+++ src/expand.rs
@@ -169,8 +169,8 @@ pub unsafe fn swap_variable_buffer(
 ///
 /// `variable_buffer` must be initialized and `off` within its allocation.
 pub unsafe fn variable_buffer_byte(off: size_t) -> ::core::ffi::c_char {
-    ::core::slice::from_raw_parts(variable_buffer as *const u8, variable_buffer_length)
-        [off as usize] as ::core::ffi::c_char
+    ::core::slice::from_raw_parts(variable_buffer as *const u8, variable_buffer_length)[off]
+        as ::core::ffi::c_char
 }
 /// Write one byte into the variable expansion buffer at `off`.
 ///
@@ -180,8 +180,8 @@ pub unsafe fn variable_buffer_byte(off: size_t) -> ::core::ffi::c_char {
 ///
 /// `variable_buffer` must be initialized and `off` within its allocation.
 pub unsafe fn set_variable_buffer_byte(off: size_t, b: ::core::ffi::c_char) {
-    ::core::slice::from_raw_parts_mut(variable_buffer as *mut u8, variable_buffer_length)
-        [off as usize] = b as u8;
+    ::core::slice::from_raw_parts_mut(variable_buffer as *mut u8, variable_buffer_length)[off] =
+        b as u8;
 }
 /// # Safety
 ///
